pull out header from layout
